Rename to cardListPropTypes
