+++ src/pages/settings/Payments/TransferBalancePage.js
@@ -22,7 +22,7 @@ import CurrentWalletBalance from '../../../components/CurrentWalletBalance';
 import walletTransferPropTypes from './walletTransferPropTypes';
 import * as PaymentMethods from '../../../libs/actions/PaymentMethods';
 import * as PaymentUtils from '../../../libs/PaymentUtils';
-import cardPropTypes from '../../../components/cardPropTypes';
+import cardListPropTypes from '../../../components/cardListPropTypes';
 import userWalletPropTypes from '../../EnablePayments/userWalletPropTypes';
 import ROUTES from '../../../ROUTES';
 
@@ -46,7 +46,7 @@ const propTypes = {
     })),
 
     /** List of card objects */
-    cardList: PropTypes.objectOf(cardPropTypes),
+    cardList: PropTypes.objectOf(cardListPropTypes),
 
     /** Wallet balance transfer props */
     walletTransfer: walletTransferPropTypes,
